Improve workers package (#26)
* Improve workers package

The workers package uses mutexes for synchronization. This approach
was put in place due to concerns related to channel performormance.
However, the use of mutexes can leak resources and requires much more
careful use of the package.

This commit replaces mutexes with channels and injects a cancellable
context when starting a worker group. This way workers will be properly
terminated when execution of a query is finished.