Handle context cancellation
The engine will not stop query execution when the query context is
cancelled. This can lead to long-running queries starting to accumulate
in the engine, which in turn can hog resources and create leaks.

This commit introduces a new Cancellable exchange operator which acts
as a middleware and checks if the context has been cancelled before
each step. The commit also makes sure worker channels are closed at the
producer side, so that we don't get "send on closed channel" errors.