Resolve panics in instant queries
The series call in instant queries can return more series than the
query can produce. This is because operations like rate will not
produce an output when a series has only one sample in the evaluation
period. The case cannot be detected during the Series call because
it requires decoding chunks, a process that happens only during Next
calls.  This can lead to the series being returned by the Series call,
but never having samples produced by a Next call.

This commit resolves the issue by explicitly checking whether a
series has produced a sample before trying to access the sample.
An illustrative test case is also added in the commit.