Optimize vector selector
We currently rely on using the MemoizedSeriesIterator from Prometheus
which can keep the last seen sample for each series. This iterator
seems to be doing many more things that are not really relevant for
decoding chunks.

This commit reimplements the logic into the vectorScanner directly
and manages to achieve a slight performance improvement by doing that.