Simplify coalesce operator
The Coalesce operator uses a RW lock in various places to merge outputs
from different operators. Even though it tries to keep critical sections
as small as possible, we still end up with the mutex in a locked state
the entire time.

This commit simplifies the locking mechanism so that the lock is acquired
and released only in one place.