Update prometheus/prometheus to v0.40.1 (#125)
This commit updates the prometheus/prometheus dependency to v0.40.1
which adds support for native histograms in TSDB and PromQL.
The commit also unblocks updating this dependency in upstream Thanos.

Since this engine does not yet support querying native histograms,
those cases will return an error for the time being. Unfortunately
the error has to be emitted during query execution since the
histogram_quantile function is now overloaded to work with both native
and counter-based histograms.

Support for native-histograms in this engine will be added as a follow up.