Batch output in step-invariant operator
The step invariant operator copies an input vector and replicates it
over the entire query range. In the current implementation, the output
of this operator is not chunked into batches, but rather expanded completely
into the total number of steps.

This leads to the binary operator not working properly since one operand
could produce 10 steps (the current batch size), and the other one
could produce steps for the entire range.

This commit changes the step invariant operator to produce output vectors
that are capped to the stepsBatch constant which is used in every other operator.