Fix instant query result sort for topk/bottomk (#156)
There is an implicit expectation that topk and bottomk instant queries
should return values in sorted order. When a grouping is used for these
aggregations, results should be sorted within individual groups.

The prometheus engine does not sort the groups themselves, which makes
it hard to do exact comparisons. However, we can now compare results
directly for these aggregations when no grouping is used.