Implement distributed deduplication (#151)
* Implement distributed deduplication

This commit adds support for deduplicating series/samples and applies
it to the distributed query engine. This is done by implementing a
Deduplication operator that removes samples with the same sample IDs from
individual step vectors. Samples are deduplicated by a
last-sample-wins strategy.

In order to properly prioritize non-stale samples in favor of stale ones,
remote engines are sorted by their max time. This way if samples come
from both older and newer engines, the ones from newer engines will be
preserved. Because of thism, stale samples will be deduplicated in favour
of non-stale ones, but will be preserved if non-stale samples are not
present inside a step vector.

* Add license header

* Fix logical tests

* Remove implicit dependency between logical plan and execution

* Add dedup comment