Fix scalar and noArg functions
Looking at PR #185, I noticed that the functions which produce nil
labels are not properly handled by the engine. They are handled
as an exception in the engine and unwinded separately. This code
has a bug where it only takes the last batch that comes out of the operator.

The reason why this was not caught by tests is because our default query
range is very short and is covered by a single batch of 10 steps.
This PR increases the default range to help uncover the issues. It also
fixes the  tests that fail due to the increased range.