Fix grouping by metric name
The labels.HashWithoutLabels function from Prometheus does not take into
account the __name__ label and explicitly excludes it from the input
into the hashing function. This leads to grouping expressions by __name__
to produce an error message in the form of vector cannot contain metrics with the same labelset.
It also interferes with deduplication in the distributed engine since
series with different __name__ labels can be treated as duplicates of
each other.

This commit replaces the use of HashWithoutLabels with BytesWithoutLabels
in the places where the engine is affected.