Add support for distributed absent
The current way to distribute functions is to run them against each engine
and concatenate the result.
The absent and absent_over_time functions are special in the sense
that they act on the entire step vector instead of on individual series.
Because of this, these two functions can currently return false positives
if a series is absent in one engine but not in another.

This commit adds special handling for the two functions that is similar to
how aggregations are distributed. Absent will be executed against each
engine as before, but all results in the root engine will be merged with a
multiplication operator. As a result, a series will be considered absent
only if it is absent in all engines, which is in line with the semantics
of the functions.