Add label based pruning in distributed optimizer (#220)
* Add label based pruning in distributed optimizer

The distributed optimizer currently creates a remote query for each
available engine. This can increase latency for queries that can be
scoped to a subset of engines based on the external labels announced
by each engine.

This commit adds support for pruning engines based on their announced
external labels. If the selectors for a query do not match any external
engine, the query will not be executed at all.

This commit also modifies the distributed optimizer to distribute binary
expressions where one operand is a constant value.

* Always match engines without external labels

* Remove obsolete comment