Add FallbackEngine to engine options (#228)
Updates the compatibility engine constructor to create an
engine, using the fallback engine provided in the options.

Since promql-engine is in early days and doesnt have any major
versions, it is finalized to use orignal function and add engine
to options.

Changes `prom` in `compatibilityEngine` to `v1.QueryEngine` type
from `promql.Engine`. Since, prom is an fallback engine and any
`QueryEngine` must work with in prom. Also, when thanos creates
a engine it get type casted into `v1.QueryEngine`, that can't
be passed to this constructor as an fallback engine option.

Signed-off-by: Pradyumna Krishna <git@onpy.in>