distribute: Add offset to engines to account for range selectors (#246)
Range selectors in PromQL need to select points from before the start of the range
query. When running the distributed optimizer, certain engines might not have sufficient
range in the past to correctly calculate the result for certain steps in the range.

To avoid returning incorrect results, this commit calculates the offset that needs to be
added to the beginning of a query for each individual remote engine. This PR also adds
detection for whether a sufficient time-based overlap exists between engines for queries
to be safely distributed. If the overlap is not sufficient, the distributed optimizer will
not modify the query and the engine will execute it in memory.