Time consumed by operator when calling Next
the TimingOperator struct wraps another VectorOperator and includes a startTime field to track the start time of each Next method call. Inside the Next method implementation of TimingOperator, it captures the current time using time.Now() and stores it in the startTime field. Then, it calls the Next method of the wrapped operator.

Signed-off-by: nishchay-veer <nishchayveer19@gmail.com>