execution: add timing information
Let's try using embedding to capture observability information for each
operator instead of wrapping operator inside of operator. Main reasoning
for this approach instead of the other way is to be able to catch more
granular data like memory allocations. With wrapping we would have to
have cross-references between operators and this seems cleaner to me.

Need to explore this further to see how it looks like.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>