Add support for injecting custom operators
This commit adds support for injecting custom operators into the engine.
The machanism involves creating a custom optimizer which produces a
logical node implementing the UserDefinedExpr interface.

This interface has a MakeExecutionOperator method which the user
can leverage to define how the execution operator is created
for the logical node.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>