Add batched streaming aggregations
With the current model we expect each Next call to return samples
for unique steps. This approach works well because of its simplicity,
but for high cardinality queries (100K+ series), it tends to use a lot
of memory because the buffers for each step tend to be big.

This commit resolves that by allowing the aggregate to handle batches
from the same step coming from subsequent Next calls. Selectors are expanded
with a batchSize parameter which can be injected when a streaming aggregate is
present in the plan. Using this parameter then can put an upper limit on
the size of the output vectors they produce.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>