Emit error when querying native histograms with extended functions
We saw panics on our setup because users tried to use the extended functions with native histograms, which led to rates calculated with only one sample in some windows.

In this PR I am introducing a safeguard against this type of situation and emiting a warn when users try to query native histograms with extended functions.

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>

Fixing test case

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>

removing dead code

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>

Return error instead of warn

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>

fixing linting

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>