Optimize aggregate accumulators
Scalar tables used for the hash aggregate with grouping labels are
currently very wasteful because of the dynamic functions they create
for each output group. Each function ends up as a new object on the heap,
so for queries where output groups have high cardinality we end up using
a lot of memory.

This commit fixes that by converting accumulators to structs, so that
functions between accumulator instances do not end up as heap-allocated objects.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>