Distribute subqueries (#370)
* Distribute subqueries

Distributing subqueries can be done in the same manner like we can
distribute regular matrix selectors. The only difference is that we
need to make sure a remote engine has sufficient scope to calculate
both the inner and outer query range. To do this, we simply add the
two ranges together when calculating the offset for the remote query
and based on this we can decide when it is safe to distribute.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Add logical plan tests

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Add test for multiple subqueries

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Fix failing test

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

---------

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>