Allow injecting custom scanners
The current implementation of the engine is coupled to the Prometheus storage
model. The input is a SeriesSet which gets iterated on one series at a time.
In cases where data is pulled from a different storage, it first has to be
transformed into the Prometheus model before it can be queried. This transformation
can be resource intensive if the data is not sorted by series and time.

This commit allows extending the engine with custom scanners which take in the
logical node and return an execution operator. With this extension point we should
get decoupling from Prometheus and can use the engine more efficiently with different
storages.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>