Fix resizing rate buffer tail
We use a single slice to rotate buffers for matrix selectors in order to
reuse memory. Sizing the slice is done using slices.Grow, but this method
does not ensure exact capacity, it only guarantees that the slice will have
at least the requested capacity.

To fix the issue, we need to use a subslice explicitly.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>