Use ringbuffer for windowing functions (#394)
* Use ringbuffer for windowing functions

The current implementation for matrix selector is based on the one from
Prometheus and uses two buffers to keep track of samples, a buffered iterator
and a slice of decoded samples for each series. In addition to being
harder to reason about, it is also inefficient because we traverse both
buffers twice for each step.

This commit rewrites the way points are selected for each step by using the
ringbuffer implemented for the subquery operator. This way we only keep a
single buffer and we don't need to manage the complexity of rotating samples
in multiple places.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Fix lint

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

---------

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>