Decouple plan from parser.Expr
The logicalplan.Node interface currently embeds parser.Expr for
backwards compatibility with the rest of the codebase. This makes it possible
to mix AST types with logical plan types which could lead to bugs in type switches
and other similiar unexpected behavior.

This commit makes sure that these two interfaces are not interchangeable anymore,
which should help us fully decouple the logical plan from the AST coming from the parser.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>