Implement json.Marshaler and json.Unmarshaler for logical plan.
We would like to take advantage of features like projections and read only
a subset of labels from store APIs when a query includes an aggregation.

In order to enable this, we need to be able to send the entire plan over the wire
since PromQL has no concept of projections. This commit implements the
json marshaler methods so that we can serialize the plan and send it in a
query request.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>