Minor fixes to InstantQueryFromPlan
The query type for instant queries from an existing plan were set to
RangeQuery because we were using the same function as the one used
for range queries. This leads to empty results for those queries.

We still need to infer the sort order for the instant query, but this
might not be needed since the sort would be done at the root anyway.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>