plan,engine: add partition labels (#569)
Add a "PartitionLabelSets" method to remote engine api. This makes sense
since a remote engine can have a big labelset but we only need a subset
of those to actually perform a partition of all engines. Operating on a
smaller labelset there helps us pushing down more expressions, since we
know if we lose one of the unimportant labels from the labelset we can
still keep pushing down.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>