ringbuffer: simplify interface (#636)
The Len() and PushIntoLast() methods of the ringbuffer interface were
redundant. Len() was only used to check if the buffer is empty, we did
just implement a convenience function that implements this logic using
the already present MaxT method and we adjusted the scanner action to
address the extlookback logic.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>