execution: cleanup warnings in vector binops
Warnings in binops have the same effect always, they get added to the
global annotations for the query and stop the computation for the
particular series.
At the same time no error that a binop can return would stop the full
flow of the query.
This indicates that we can communicate warnings through the returned
error in the binop and just add them to the annotations and continue the
flow. This simplifies the annotation handling in vector binary
operations a little bit.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>